This code is a sample that implements many school management aspects; it serves as a good starting point for developers
This is a C# desktop application that connects to Azure SQL DB to get
student info. It is also possible to read student info from a text file.
The code is not functional; the user must first write the name of his
Azure DB, server, and password in the appropriate .cs files